requirements.txt
tweepy
nltk
spellchecker
pyspellchecker
python-dateutil
textblob
cloudpickle
pandas
matplotlib
sklearn
